Created new fragment structure, with nav graph and a NavGraphHost. MusicService is now the only Acitivity running Libtorrent, the rest of UI and interactions happen on Fragments.
