+++ musicdao/build.gradle
@@ -2,6 +2,7 @@ apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'org.jlleitschuh.gradle.ktlint'
+apply plugin: 'androidx.navigation.safeargs.kotlin'
 
 android {
     compileSdkVersion 29
@@ -60,6 +61,7 @@ dependencies {
     implementation 'androidx.navigation:navigation-fragment-ktx:2.2.2'
     implementation 'androidx.navigation:navigation-ui-ktx:2.2.2'
     implementation 'com.turn:ttorrent-core:1.5'
+    implementation 'com.google.android:flexbox:2.0.1'
     implementation files('libs/jlibtorrent-1.2.5.0.jar')
     implementation files('libs/jlibtorrent-android-arm64-1.2.5.0.jar')
     implementation files('libs/jlibtorrent-android-arm-1.2.5.0.jar')
