--- musicdao/src/main/java/com/example/musicdao/MainFragment.kt
@@ -1,11 +1,0 @@
-package com.example.musicdao
-
-import androidx.fragment.app.Fragment
-import nl.tudelft.trustchain.common.ui.BaseFragment
-
-/**
- * A simple [Fragment] subclass.
- * Use the [MainFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-class MainFragment : BaseFragment(R.layout.music_app_main)
