+++ musicdao/src/main/java/com/example/musicdao/MusicService.kt
@@ -1,20 +1,19 @@
 package com.example.musicdao
 
-import android.content.Intent
+//import kotlinx.android.synthetic.main.music_app_main.*
 import android.os.Bundle
-import android.text.Editable
 import android.util.Log
 import android.view.MenuItem
 import android.widget.Toast
 import androidx.core.app.NavUtils
 import androidx.preference.PreferenceManager
-import com.example.musicdao.ui.SubmitReleaseDialog
 import com.frostwire.jlibtorrent.FileStorage
-import kotlinx.android.synthetic.main.fragment_release.*
-import kotlinx.android.synthetic.main.music_app_main.*
 import nl.tudelft.ipv8.android.IPv8Android
 import nl.tudelft.ipv8.android.keyvault.AndroidCryptoProvider
-import nl.tudelft.ipv8.attestation.trustchain.*
+import nl.tudelft.ipv8.attestation.trustchain.BlockListener
+import nl.tudelft.ipv8.attestation.trustchain.BlockSigner
+import nl.tudelft.ipv8.attestation.trustchain.TrustChainBlock
+import nl.tudelft.ipv8.attestation.trustchain.TrustChainCommunity
 import nl.tudelft.ipv8.keyvault.PrivateKey
 import nl.tudelft.ipv8.util.hexToBytes
 import nl.tudelft.ipv8.util.toHex
@@ -25,10 +24,8 @@ const val PREPARE_SIZE_KB: Long = 10 * 512L
 
 class MusicService : BaseActivity() {
     private var currentMagnetLoading: String? = null
-    private val defaultTorrent =
-        "magnet:?xt=urn:btih:2803173609ad794d2789da6a6852fc1dbda7b7bf&dn=tru1992-07-23"
-    private val trackLibrary: TrackLibrary =
-        TrackLibrary(this)
+    public val trackLibrary: TrackLibrary =
+        TrackLibrary()
 
     override val navigationGraph = R.navigation.musicdao_navgraph
 
@@ -45,7 +42,6 @@ class MusicService : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         clearCache()
-        setContentView(R.layout.music_app_main)
         supportActionBar?.title = "Music app"
 
         registerBlockListener()
@@ -56,18 +52,9 @@ class MusicService : BaseActivity() {
 //            mainLinearLayout.addView(AudioPlayer.getInstance(this), 0)
 //        }
 
-        torrentButton.setOnClickListener {
-            createDefaultBlock()
-        }
-
-        shareTrackButton.setOnClickListener {
-            selectLocalTrackFile()
-        }
-
         Thread(Runnable {
             trackLibrary.startDHT()
         }).start()
-        iterateClientConnectivity()
     }
 
     fun prepareNextTrack() {
@@ -88,85 +75,6 @@ class MusicService : BaseActivity() {
                 it.deleteRecursively()
             }
         }
-    }
-
-    /**
-     * Iteratively update and show torrent client connectivity
-     */
-    private fun iterateClientConnectivity() {
-        val thread: Thread = object : Thread() {
-            override fun run() {
-                try {
-                    while (!this.isInterrupted) {
-                        sleep(1000)
-                        runOnUiThread {
-                            val text =
-                                "UP: ${trackLibrary.getUploadRate()} DOWN: ${trackLibrary.getDownloadRate()} DHT: ${trackLibrary.getDhtNodes()}"
-                            torrentClientInfo.text = text
-                        }
-                    }
-                } catch (e: Exception) {
-                }
-            }
-        }
-        thread.start()
-    }
-
-    /**
-     * Select an audio file from local disk
-     */
-    private fun selectLocalTrackFile() {
-        val chooseFile = Intent(Intent.ACTION_GET_CONTENT)
-        chooseFile.type = "audio/*"
-        val chooseFileActivity = Intent.createChooser(chooseFile, "Choose a file")
-        startActivityForResult(chooseFileActivity, 1)
-        val uri = chooseFileActivity.data
-        if (uri != null) {
-            println(uri.path)
-        }
-    }
-
-    /**
-     * This is called when the chooseFile is completed
-     */
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        val uri = data?.data
-        if (uri != null) {
-            // This should be reached when the chooseFile intent is completed and the user selected
-            // an audio file
-            val magnet = trackLibrary.seedFile(applicationContext, uri)
-            publishTrack(magnet)
-        }
-    }
-
-    /**
-     * Once a magnet link to publish is chosen, show an alert dialog which asks to add metadata for
-     * the Release (album title, release date etc)
-     */
-    private fun publishTrack(magnet: String) {
-        this.currentMagnetLoading = magnet
-        SubmitReleaseDialog(this)
-            .show(supportFragmentManager, "Submit metadata")
-    }
-
-    /**
-     * After the user inserts some metadata for the release to be published, this function is called
-     * to create the proposal block
-     */
-    fun finishPublishing(title: Editable?, artists: Editable?, releaseDate: Editable?) {
-        val myPeer = IPv8Android.getInstance().myPeer
-
-        val transaction = mapOf(
-            "publisher" to myPeer.mid,
-            "magnet" to currentMagnetLoading,
-            "title" to title.toString(),
-            "artists" to artists.toString(),
-            "date" to releaseDate.toString()
-        )
-        val trustchain = IPv8Android.getInstance().getOverlay<TrustChainCommunity>()!!
-        Toast.makeText(applicationContext, "Creating proposal block", Toast.LENGTH_SHORT).show()
-        trustchain.createProposalBlock("publish_release", transaction, myPeer.publicKey.keyToBin())
     }
 
     private fun registerBlockSigner() {
@@ -212,14 +120,6 @@ class MusicService : BaseActivity() {
                 Log.d("TrustChainDemo", "onBlockReceived: ${block.blockId} ${block.transaction}")
             }
         })
-    }
-
-    /**
-     * Creates a trustchain block which uses the example creative commons release magnet
-     * This is useful for testing the download speed of tracks over libtorrent
-     */
-    private fun createDefaultBlock() {
-        publishTrack(defaultTorrent)
     }
 
     companion object {
