+++ musicdao/src/main/java/com/example/musicdao/ReleaseOverviewFragment.kt
@@ -1,6 +1,13 @@
 package com.example.musicdao
 
-import android.app.Activity
+import android.os.Bundle
+import android.view.View
+import androidx.navigation.fragment.findNavController
 import nl.tudelft.trustchain.common.ui.BaseFragment
 
-class ReleaseOverviewFragment: BaseFragment(R.layout.release_overview_fragment)
+class ReleaseOverviewFragment: BaseFragment(R.layout.fragment_release_overview) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+}
