+++ musicdao/src/main/java/com/example/musicdao/ui/SubmitReleaseDialog.kt
@@ -7,9 +7,10 @@ import android.os.Bundle
 import android.widget.EditText
 import androidx.fragment.app.DialogFragment
 import com.example.musicdao.MusicService
+import com.example.musicdao.PlaylistFragment
 import com.example.musicdao.R
 
-class SubmitReleaseDialog(private val musicService: MusicService) : DialogFragment() {
+class SubmitReleaseDialog(private val playlistFragment: PlaylistFragment) : DialogFragment() {
     private lateinit var title: EditText
     private lateinit var artists: EditText
     private lateinit var releaseDate: EditText
@@ -29,7 +30,7 @@ class SubmitReleaseDialog(private val musicService: MusicService) : DialogFragme
                         val artistEditText = dialog?.findViewById<EditText>(R.id.artists)
                         val releaseDateEditText =
                             dialog?.findViewById<EditText>(R.id.release_date)
-                        musicService.finishPublishing(
+                        playlistFragment.finishPublishing(
                             titleEditText?.text,
                             artistEditText?.text,
                             releaseDateEditText?.text
