+++ musicdao/src/main/res/layout/fragment_release_overview.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<com.google.android.flexbox.FlexboxLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    app:flexWrap="wrap"
+    android:background="@color/colorPrimaryPlayer">
+
+    <fragment
+        android:id="@+id/release_cover_fragment1"
+        class="com.example.musicdao.ReleaseCoverFragment"
+        android:layout_width="150dp"
+        android:layout_height="150dp"
+        android:layout_margin="@dimen/fab_margin" />
+
+    <fragment
+        android:id="@+id/release_cover_fragment2"
+        class="com.example.musicdao.ReleaseCoverFragment"
+        android:layout_width="150dp"
+        android:layout_height="150dp"
+        android:layout_margin="@dimen/fab_margin" />
+
+    <fragment
+        android:id="@+id/release_cover_fragment3"
+        class="com.example.musicdao.ReleaseCoverFragment"
+        android:layout_width="150dp"
+        android:layout_height="150dp"
+        android:layout_margin="@dimen/fab_margin" />
+
+</com.google.android.flexbox.FlexboxLayout>
