+++ musicdao/src/main/res/layout/music_app_main.xml
@@ -4,7 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainFragment"
     android:background="#303030"
     android:orientation="vertical">
